+++ api-derive/bagsList/types.ts
@@ -1,8 +1,8 @@
 // Copyright 2017-2022 @polkadot/api-derive authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import type { PalletBagsListListBag, PalletBagsListListNode } from 'https://deno.land/x/polkadot@0.0.4/types/lookup.ts';
-import type { BN } from 'https://deno.land/x/polkadot@0.0.4/util/mod.ts';
+import type { PalletBagsListListBag, PalletBagsListListNode } from 'https://deno.land/x/polkadot/types/lookup.ts';
+import type { BN } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 export interface Bag {
   bag: PalletBagsListListBag | null;