+++ api/promise/Api.ts
@@ -3,7 +3,7 @@
 
 import type { ApiOptions, UnsubscribePromise } from '../types/index.ts';
 
-import { objectSpread } from 'https://deno.land/x/polkadot@0.0.4/util/mod.ts';
+import { objectSpread } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 import { ApiBase } from '../base/index.ts';
 import { Combinator, CombinatorCallback, CombinatorFunction } from './Combinator.ts';
@@ -28,7 +28,7 @@ import { promiseTracker, toPromiseMethod } from './decorateMethod.ts';
  * <BR>
  *
  * ```javascript
- * import ApiPromise from 'https://deno.land/x/polkadot@0.0.4/api/promise/index.ts';
+ * import ApiPromise from 'https://deno.land/x/polkadot/api/promise/index.ts';
  *
  * // initialise via static create
  * const api = await ApiPromise.create();
@@ -44,7 +44,7 @@ import { promiseTracker, toPromiseMethod } from './decorateMethod.ts';
  * <BR>
  *
  * ```javascript
- * import { ApiPromise, WsProvider } from 'https://deno.land/x/polkadot@0.0.4/api/mod.ts';
+ * import { ApiPromise, WsProvider } from 'https://deno.land/x/polkadot/api/mod.ts';
  *
  * // initialise a provider with a specific endpoint
  * const provider = new WsProvider('wss://example.com:9944')
@@ -72,7 +72,7 @@ import { promiseTracker, toPromiseMethod } from './decorateMethod.ts';
  * <BR>
  *
  * ```javascript
- * import ApiPromise from 'https://deno.land/x/polkadot@0.0.4/api/promise/index.ts';
+ * import ApiPromise from 'https://deno.land/x/polkadot/api/promise/index.ts';
  *
  * ApiPromise.create().then((api) => {
  *   const [nonce] = await api.query.system.account(keyring.alice.address);
@@ -105,7 +105,7 @@ export class ApiPromise extends ApiBase<'promise'> {
    * <BR>
    *
    * ```javascript
-   * import Api from 'https://deno.land/x/polkadot@0.0.4/api/promise/index.ts';
+   * import Api from 'https://deno.land/x/polkadot/api/promise/index.ts';
    *
    * new Api().isReady.then((api) => {
    *   api.rpc.subscribeNewHeads((header) => {
@@ -137,7 +137,7 @@ export class ApiPromise extends ApiBase<'promise'> {
    * <BR>
    *
    * ```javascript
-   * import Api from 'https://deno.land/x/polkadot@0.0.4/api/promise/index.ts';
+   * import Api from 'https://deno.land/x/polkadot/api/promise/index.ts';
    *
    * Api.create().then(async (api) => {
    *   const timestamp = await api.query.timestamp.now();