+++ rpc-provider/mock/types.ts
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import type { Server } from 'https://esm.sh/mock-socket@9.1.5';
-import type { Constructor } from 'https://deno.land/x/polkadot@0.0.4/types/types/index.ts';
+import type { Constructor } from 'https://deno.land/x/polkadot/types/types/index.ts';
 
 // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
 export interface Global extends NodeJS.Global {
@@ -11,7 +11,7 @@ export interface Global extends NodeJS.Global {
 }
 
 export interface Mock {
-  body: Record<string, any>;
+  body: Record<string, Record<string, unknown>>;
   requests: number;
   server: Server;
   done: () => Record<string, unknown>;