+++ types-codec/native/Date.ts
@@ -1,12 +1,13 @@
 // Copyright 2017-2022 @polkadot/types-codec authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import type { HexString } from 'https://deno.land/x/polkadot@0.0.4/util/types.ts';
+import type { HexString } from 'https://deno.land/x/polkadot/util/types.ts';
 import type { AnyNumber, Inspect, INumber, IU8a, Registry, UIntBitLength } from '../types/index.ts';
 
-import { BN, bnToBn, bnToHex, bnToU8a, isString, isU8a, u8aToBn } from 'https://deno.land/x/polkadot@0.0.4/util/mod.ts';
+import { BN, bnToBn, bnToHex, bnToU8a, isString, isU8a, u8aToBn } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 const BITLENGTH: UIntBitLength = 64;
+const U8A_OPTS = { bitLength: BITLENGTH, isLe: true };
 
 function decodeDate (value: CodecDate | Date | AnyNumber): Date {
   if (isU8a(value)) {
@@ -156,6 +157,6 @@ export class CodecDate extends Date implements INumber {
    */
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   public toU8a (isBare?: boolean): Uint8Array {
-    return bnToU8a(this.toNumber(), BITLENGTH, true);
+    return bnToU8a(this.toNumber(), U8A_OPTS);
   }
 }