+++ types/interfaces/imOnline/types.ts
@@ -1,10 +1,10 @@
 // Auto-generated via `yarn polkadot-types-from-defs`, do not edit
 /* eslint-disable */
 
-import type { Bytes, Struct, Vec, u32 } from 'https://deno.land/x/polkadot@0.0.4/types-codec/mod.ts';
-import type { Signature } from 'https://deno.land/x/polkadot@0.0.4/types/interfaces/extrinsics/index.ts';
-import type { BlockNumber } from 'https://deno.land/x/polkadot@0.0.4/types/interfaces/runtime/index.ts';
-import type { SessionIndex } from 'https://deno.land/x/polkadot@0.0.4/types/interfaces/session/index.ts';
+import type { Bytes, Struct, Vec, WrapperOpaque, u32 } from 'https://deno.land/x/polkadot/types-codec/mod.ts';
+import type { Signature } from 'https://deno.land/x/polkadot/types/interfaces/extrinsics/index.ts';
+import type { BlockNumber } from 'https://deno.land/x/polkadot/types/interfaces/runtime/index.ts';
+import type { SessionIndex } from 'https://deno.land/x/polkadot/types/interfaces/session/index.ts';
 
 /** @name AuthIndex */
 export interface AuthIndex extends u32 {}
@@ -30,7 +30,7 @@ export interface HeartbeatTo244 extends Struct {
 }
 
 /** @name OpaqueMultiaddr */
-export interface OpaqueMultiaddr extends Bytes {}
+export interface OpaqueMultiaddr extends WrapperOpaque<Bytes> {}
 
 /** @name OpaqueNetworkState */
 export interface OpaqueNetworkState extends Struct {
@@ -39,6 +39,6 @@ export interface OpaqueNetworkState extends Struct {
 }
 
 /** @name OpaquePeerId */
-export interface OpaquePeerId extends Bytes {}
+export interface OpaquePeerId extends WrapperOpaque<Bytes> {}
 
 export type PHANTOM_IMONLINE = 'imOnline';